fetch data from localhost
