+++ client/src/components/molecules/Classes.jsx
@@ -3,49 +3,41 @@ import ClassRow from "./ClassRow";
 
 const Classes = () => {
   const [data, setData] = useState([]);
-  // useEffect(() => {
-  //   fetch("https://web-wanderers-cyf.herokuapp.com/jobs")
-  //     .then((res) => res.json())
-  //     .then((data) => setData(data));
-  // }, []);
-
-  // const filteredArray = [];
-
-  // data.map((item) => {
-  //   filteredArray.push(
-  //     (({ date, availableSpots }) => ({ date, availableSpots }))(item)
-  //   );
-  // });
-  // const sumObject = {};
-
-  // filteredArray.map((item) => {
-  //   if (sumObject.hasOwnProperty(item.date)) {
-  //     sumObject[item.date] =
-  //       parseInt(sumObject[item.date]) + parseInt(item.availableSpots);
-  //   } else {
-  //     sumObject[item.date] = item.availableSpots;
-  //   }
-  // });
-  // const sumArray = [];
-
-  // for (let key in sumObject)
-  //   sumArray.push({ date: key, availableSpots: sumObject[key] });
-
-  const getData = async () => {
-    const res = await fetch("https://web-wanderers-cyf.herokuapp.com/jobs");
-    const data = await res.json();
-    setData(data);
-  };
-  console.log(data);
   useEffect(() => {
-    getData();
+    fetch("http://localhost:4000/jobs")
+      .then((res) => res.json())
+      .then((data) => setData(data));
   }, []);
+
+  const filteredArray = [];
+
+  data.map((item) => {
+    filteredArray.push(
+      (({ date, availableSpots }) => ({ date, availableSpots }))(item)
+    );
+  });
+  const sumObject = {};
+
+  filteredArray.map((item) => {
+    if (sumObject.hasOwnProperty(item.date)) {
+      sumObject[item.date] =
+        parseInt(sumObject[item.date]) + parseInt(item.availableSpots);
+    } else {
+      sumObject[item.date] = item.availableSpots;
+    }
+  });
+  const sumArray = [];
+
+  for (let key in sumObject)
+    sumArray.push({ date: key, availableSpots: sumObject[key] });
+
+  console.log(data);
 
   return (
     <>
       <h2 class="text-center fw-bold">Availabe Dates and Slots</h2>
       <div class="row d-flex justify-content-center m-3 mb-5">
-        {data.map((oneClass, index) => (
+        {sumArray.map((oneClass, index) => (
           <ClassRow
             key={index}
             available={oneClass.availableSpots}
