+++ client/src/components/pages/Contact.jsx
@@ -4,9 +4,14 @@ const Contact = () => {
   return (
     <div className="m-2 p-4 border text-center">
       <h2>Akwaaba</h2>
-      <p>Princess May Primary School <br /> Barrett's Grove Entrance (off Stoke Newington Road) <br /> N16 8AJ</p>
+      <p>
+        Princess May Primary School <br /> Barrett's Grove Entrance (off Stoke
+        Newington Road) <br /> N16 8AJ
+      </p>
 
-      <p className="mt-4">Please send us email and follow us on social media from the link below</p>
+      <p className="mt-4">
+        Please send us email and follow us on social media from the link below
+      </p>
     </div>
   );
 };
